# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Soccer Connect API",
    "description": "Backend API for Soccer Connect - GTA soccer community platform",
    "version": "1.0.0",
    "contact": {
      "name": "Soccer Connect Team"
    }
  },
  "servers": [
    {
      "url": "https://soccer-connect-api.vercel.app/api",
      "description": "Production"
    },
    {
      "url": "http://localhost:8046/api",
      "description": "Local Development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
...
  "paths": {
...
    "/users": {
      "get": {
        "tags": ["Users"],
        "summary": "Get all users with pagination and filters",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20 } },
          { "name": "position", "in": "query", "schema": { "type": "string", "enum": ["goalkeeper", "defender", "midfielder", "forward"] } },
          { "name": "search", "in": "query", "schema": { "type": "string" } },
          { "name": "available", "in": "query", "schema": { "type": "string", "enum": ["true", "false"] } }
        ],
        "responses": {
          "200": { "description": "List of users with pagination" },
          "401": { "description": "Not authorized" }
        }
      }
    },
...
    "/events/{id}": {
      "get": {
        "tags": ["Events"],
        "summary": "Get single event",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Event details" },
          "404": { "description": "Event not found" }
        }
      },
      "put": {
        "tags": ["Events"],
        "summary": "Update event",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Event updated" },
          "403": { "description": "Not authorized" }
        }
      },
      "delete": {
        "tags": ["Events"],
        "summary": "Delete event",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Event deleted" }
        }
      }
    },
...
    "/teams": {
      "get": {
        "tags": ["Teams"],
        "summary": "Get all teams with filters and pagination",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20 } },
          { "name": "skill_level", "in": "query", "schema": { "type": "string" } },
          { "name": "recruiting", "in": "query", "schema": { "type": "string", "enum": ["true", "false"] } },
          { "name": "search", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "List of teams" }
        }
      },
      "post": {
        "tags": ["Teams"],
        "summary": "Create a new team",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["team_name"],
                "properties": {
                  "team_name": { "type": "string" },
                  "description": { "type": "string" },
                  "skill_level": { "type": "string", "enum": ["recreational", "intermediate", "competitive"] },
                  "location": { "type": "object" },
                  "max_members": { "type": "integer" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Team created" }
        }
      }
    },
...
    "/teams/{id}": {
      "get": {
        "tags": ["Teams"],
        "summary": "Get single team",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Team details" },
          "404": { "description": "Team not found" }
        }
      },
      "put": {
        "tags": ["Teams"],
        "summary": "Update team",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Team updated" }
        }
      },
      "delete": {
        "tags": ["Teams"],
        "summary": "Delete team",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Team deleted" }
        }
      }
    },
...
    "/classifieds/{id}": {
      "get": {
        "tags": ["Classifieds"],
        "summary": "Get single classified",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Classified details" }
        }
      },
      "put": {
        "tags": ["Classifieds"],
        "summary": "Update classified",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Classified updated" }
        }
      },
      "delete": {
        "tags": ["Classifieds"],
        "summary": "Delete classified",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Classified deleted" }
        }
      }
    },
...
    "/admin/stats": {
      "get": {
        "tags": ["Admin"],
        "summary": "Get admin dashboard statistics",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Dashboard stats" }
        }
      }
    },
    "/admin/events/pending": {
      "get": {
        "tags": ["Admin"],
        "summary": "Get events pending approval",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Pending events" }
        }
      }
    },
    "/admin/events/{id}/approve": {
      "put": {
        "tags": ["Admin"],
        "summary": "Approve an event",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Event approved" }
        }
      }
    },
    "/admin/events/{id}/reject": {
      "put": {
        "tags": ["Admin"],
        "summary": "Reject an event",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Event rejected" }
        }
      }
    },
    "/admin/teams/pending": {
      "get": {
        "tags": ["Admin"],
        "summary": "Get teams pending approval",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Pending teams" }
        }
      }
    },
    "/admin/teams/{id}/approve": {
      "put": {
        "tags": ["Admin"],
        "summary": "Approve a team",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Team approved" }
        }
      }
    },
    "/admin/teams/{id}/reject": {
      "put": {
        "tags": ["Admin"],
        "summary": "Reject a team",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Team rejected" }
        }
      }
    },
    "/admin/classifieds/pending": {
      "get": {
        "tags": ["Admin"],
        "summary": "Get classifieds pending approval",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Pending classifieds" }
        }
      }
    },
    "/admin/classifieds/{id}/approve": {
      "put": {
        "tags": ["Admin"],
        "summary": "Approve a classified",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Classified approved" }
        }
      }
    },
    "/admin/classifieds/{id}/reject": {
      "put": {
        "tags": ["Admin"],
        "summary": "Reject a classified",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Classified rejected" }
        }
      }
    },
    "/admin/teams": {
      "get": {
        "tags": ["Admin"],
        "summary": "Get all teams with filters (admin)",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "All teams" }
        }
      }
    },
    "/admin/teams/{id}": {
      "put": {
        "tags": ["Admin"],
        "summary": "Update team details (admin)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Team updated" }
        }
      },
      "delete": {
        "tags": ["Admin"],
        "summary": "Delete a team (admin)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Team deleted" }
        }
      }
    },
    "/admin/users": {
      "get": {
        "tags": ["Admin"],
        "summary": "Get all users with filters (admin)",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "All users" }
        }
      }
    },
    "/admin/users/{id}/role": {
      "put": {
        "tags": ["Admin"],
        "summary": "Change user role",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["user_type"],
                "properties": {
                  "user_type": { "type": "string", "enum": ["player", "admin"] }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Role updated" }
        }
      }
    },
    "/admin/users/{id}/ban": {
      "put": {
        "tags": ["Admin"],
        "summary": "Ban/Unban a user",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "User ban status toggled" }
        }
      }
    }
  }
}